URL Rewriting Middleware update
