+++ aspnetcore/fundamentals/url-rewriting/sample/IISUrlRewrite.xml
