+++ aspnetcore/fundamentals/url-rewriting/sample/RewriteRules.cs
@@ -1,5 +1,4 @@
 using System;
-using System.IO;
 using System.Text.RegularExpressions;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Rewrite;
