--- aspnetcore/fundamentals/url-rewriting/samples/2.x/ApacheModRewrite.txt
@@ -1,2 +1,0 @@
-# Rewrite path with additional sub directory
-RewriteRule ^/apache-mod-rules-redirect/(.*) /redirected?id=$1 [L,R=302]
