--- aspnetcore/fundamentals/url-rewriting/samples/2.x/IISUrlRewrite.xml
@@ -1,8 +1,0 @@
-<rewrite>
-  <rules>
-    <rule name="Rewrite segment to id querystring" stopProcessing="true">
-      <match url="^iis-rules-rewrite/(.*)$" />
-      <action type="Rewrite" url="rewritten?id={R:1}" appendQueryString="false"/>
-    </rule>
-  </rules>
-</rewrite>
